Option, Quit, Search and Show moved to Options package
