+++ src/main/java/com/pelensky/contactmanager/Options/Quit.java
@@ -1,11 +1,15 @@
-package com.pelensky.contactmanager;
+package com.pelensky.contactmanager.Options;
+
+import com.pelensky.contactmanager.AppRunner;
+import com.pelensky.contactmanager.IO;
+import com.pelensky.contactmanager.Options.Option;
 
 public class Quit implements Option {
 
     private IO io;
     private AppRunner appRunner;
 
-    Quit(IO io, AppRunner appRunner) {
+    public Quit(IO io, AppRunner appRunner) {
         this.io = io;
         this.appRunner = appRunner;
     }
